requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
pymupdf>=1.23.0
docx2txt>=0.8
pdfplumber>=0.10.0
pandas>=2.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
weasyprint>=60.0
streamlit-option-menu>=0.3.0
plotly>=5.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
textblob>=0.17.0
nltk>=3.8.0
scikit-learn>=1.3.0
wordcloud>=1.9.0
faker>=19.0.0
